refactor(hooks & services): creating a utils folder for utility functions, rename file
