+++ src/hooks/useQuote.js
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { initialQuoteData, QUERY_QUOTE_KEY } from '@/constants/index';
-import fetchQuote from '@/services/quoteService';
+import fetchQuote from '@/services/utils/fetchQuote';
 
 const useQuote = () =>
   useQuery({
